requirements.txt
flask
google-generativeai
beautifulsoup4
pydub
speechrecognition
requests
yt-dlp
python-dotenv
